sync service refactoring
