+++ server/src/service/gcp-account-sync-service.js
@@ -1,6 +1,8 @@
 const accountService = require('./account-service');
 const activeDirectoryService = require('./active-directory-service');
 const awsSnsService = require('./aws-sns-service');
+const pubSubService = require('./gcp-pubsub-service');
+const { groupBy, mapValues, orderBy } = require('lodash');
 
 const tableName = process.env.ACCOUNT_GCP_TABLE;
 
@@ -47,8 +49,64 @@ const syncOwners = async () => {
   }
 }
 
+// Description
+// GCP does not have an API for getting spent budget per account
+// Instead of that, we activated budget notification (on budget threshold) to be sent to pubsub topic.
+// Lambda task: Fetching messages from topic and updating account/project 'spent amount' value in AWS Dynamodb Table
+// Lambda will retrigger new execution if there are more messages to be proceeded
+const syncBudgets = async () => {
+
+  // get message from gcp pub/sub
+  let messagesResponse = await pubSubService.getMessages();
+  const rawMessages = messagesResponse.data.receivedMessages || [];
+
+  // encode message content
+  const messages = rawMessages.map((message) => {
+    const data = Buffer.from(message.message.data, 'base64');
+    return JSON.parse(data.toString());
+  });
+
+  // sort budget per name and costAmount (possible duplicates)
+  const sortedBudgets = mapValues(groupBy(messages, message => message.budgetDisplayName), object => orderBy(object, 'costAmount', 'desc'));
+
+  // get existing accounts from dynamodb
+  const accounts = await accountService.getAccounts(tableName);
+
+  // budget name === ${account/project name} + ${suffix '-budget'}
+  const budgetNames = Object.keys(sortedBudgets);
+
+  const events = [];
+  // iterate through budget messages from pubsub and pick accounts for update
+  budgetNames.forEach(budgetName => {
+    const accountName = budgetName.replace('-budget', '');
+    const accountForUpdate = accounts.find(account => account.name === accountName);
+    // if condition is true, call dynamodb and update account actualSpend value
+    if (accountForUpdate && Number(accountForUpdate.actualSpend) < Number(sortedBudgets[budgetName][0].costAmount)) {
+      console.log(`Update budget for: ${accountForUpdate.id} - ${accountForUpdate.name}, old cost ${accountForUpdate.actualSpend}, new cost ${sortedBudgets[budgetName][0].costAmount}`)
+      // TODO call dynamo for item update, uncomment real call to dynamoDB
+      // get the biggest sorted value for actualSpend (because of message duplicates)
+      // events.push(accountService.updateAccount({ id: accountForUpdate.id, actualSpend: `${sortedBudgets[name][0].costAmount}` }, tableName));
+      events.push(Promise.resolve(accountForUpdate.name));
+    }
+  })
+  await Promise.all(events);
+
+  // purge processed messages
+  const ackIds = rawMessages.map(message => message.ackId);
+  await pubSubService.acknowledgeMessages(ackIds);
+  console.log('acknowledge messages finished');
+
+  // call for a new batch of messages
+  if (messages.length) {
+    await syncBudgets();
+  }
+  
+  return { success: true, message: 'Budget sync is done' };
+}
+
 module.exports = {
   syncAccounts,
   syncAccountsMembers,
-  syncOwners
+  syncOwners,
+  syncBudgets
 };
